activities: Let NotifyWatchers return a list of Actions
In the near future, we will wish to have access to the Actions inserted
into the `action` table, but only those that were inserted for the
owning user. To facilitate this, let `NotifyWatchers()` (and
`NotifyWatchersActions`) return these actions.

The list is currently ignored, but `services/feed` was modified to make
it easier to add other notifiers, which will use the list returned by
`NotifyWatchers`.

Signed-off-by: Gergely Nagy <forgejo@gergo.csillger.hu>